requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
comm==0.1.3
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.13
distlib==0.3.6
executing==1.2.0
feather-format==0.4.1
filelock==3.12.0
fonttools==4.39.4
identify==2.5.24
idna==3.4
ipykernel==6.24.0
ipython==8.14.0
ipywidgets==8.0.7
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
kiwisolver==1.4.4
llvmlite==0.40.1
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
nest-asyncio==1.5.6
nodeenv==1.8.0
numba==0.57.1
numpy==1.24.3
packaging==23.1
pandas==2.0.2
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
pre-commit==3.3.2
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
PyGithub==1.58.2
Pygments==2.15.1
PyJWT==2.7.0
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
requests==2.30.0
scikit-learn==1.2.2
scipy==1.11.0
shap==0.42.0
six==1.16.0
slicer==0.0.7
stack-data==0.6.2
tabulate==0.9.0
threadpoolctl==3.1.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
tzdata==2023.3
urllib3==2.0.2
virtualenv==20.23.0
wcwidth==0.2.6
widgetsnbextension==4.0.8
wrapt==1.15.0
